update: consumer for route assigned events

+++ src/main/java/vn/com/routex/hub/management/service/domain/route/model/VehicleSnapshot.java
@@ -14,6 +14,6 @@ import lombok.experimental.SuperBuilder;
 public class VehicleSnapshot {
     private String id;
     private String vehiclePlate;
-    private Integer seatCapacity;
+    private Long seatCapacity;
     private boolean hasFloor;
 }
